requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
blinker==1.8.2
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
deprecation==2.1.0
distlib==0.3.8
distro==1.9.0
Django==5.0.6
environs==14.3.0
executing==2.2.0
fastapi==0.116.1
fastapi_cors==0.0.6
filelock==3.15.4
Flask==3.0.3
flask-cors==5.0.1
fonttools==4.56.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
gotrue==2.12.3
groq==0.30.0
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
ipykernel==6.29.5
ipython==8.32.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiter==0.9.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.8
MarkupSafe==2.1.5
marshmallow==4.0.0
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
narwhals==1.44.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.2.3
openai==1.76.2
packaging==24.2
pandas==2.3.0
parso==0.8.4
pillow==11.1.0
platformdirs==4.2.2
postgrest==1.1.1
prompt_toolkit==3.0.50
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.0.0
pure_eval==0.2.3
pyarrow==20.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
realtime==2.6.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.4
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.5.3
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.0
stack-data==0.6.3
starlette==0.47.2
storage3==0.12.0
streamlit==1.46.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
sympy==1.14.0
tenacity==9.1.2
tokenizers==0.21.2
toml==0.10.2
torch==2.7.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.14.1
tzdata==2024.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
virtualenv==20.26.3
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.0.4
